third_party/blink/web_tests/platform/generic/fast/parser/document-close-iframe-load-partial-entity-expected.txt

ALERT: This test passes if the text "&g" appears below.
&g
